requirements.txt
asgiref==3.7.2
crispy-bootstrap5==0.7
django-crispy-forms==2.0
django-taggit==4.0.0
Django==4.2.5
Pillow==10.0.0
pip==22.3.1
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.7.1
wheel==0.38.4
